Create README.mdMa Africa is bless with so many resources but the problem is how to manage it. On of the cause of hunger in Africa is lack of storage, Transportation and insurgency and banditry. How to convey goods from were there is surplus to were they are needed is a problem, how to preserve them is another problem. Those who produce these goods are been murder by banditry and local herders. These are some of the challenges we face in Africa were there is lots of natural resources.
Ma Africa is bless with so many resources but the problem is how to manage it. On of the cause of hunger in Africa is lack of storage, Transportation and insurgency and banditry. How to convey goods from were there is surplus to were they are needed is a problem, how to preserve them is another problem. Those who produce these goods are been murder by banditry and local herders. These are some of the challenges we face in Africa were there is lots of natural resources.